Deployment/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
Pillow
numpy
scikit-learn==1.0.2